Assignment template URLs work.
Implements template URLs in the form of `#template=GDRIVEID`, where `GDRIVEID` is the gdrive id of a public folder containing a file named `template.arr`. The contents of `template.arr` are loaded into the editor.

Unfortunately, this is overly simplistic because, unlike usual CPO, we need to support a consistent notion of an associated assignment. For a typical programming homework, we distribute a `#share` URL, which graduates to a student-specific `#program` URL once they make changes, and the provenance of the file is forgotten. For examplar, we still want the student to be able to save their changes, but we also need to retain a notion of associated assignment so we can run their file against wheats and chaffs.