Support a notion of an associated assignment.
Resolves the shortcoming of 17aec5c54d913bd3d9faff3704416e8aa11e5701.

When a user navigates to a `#template=GDRIVE` URL, we search their gdrive for a file with the property `assignment` equal to `GDRIVE`. If such a file exists, we load it. Otherwise, the file named `template.arr` is selected from the folder of id `GDRIVE`. We copy it to the user's gdrive, set the property `assignment` of the copied file equal to `GDRIVE`, and share the copied file with a course staff google account.