Adding 'hint' functionality to Examplar (#31)
The idea at the core of this change is that wheat failing tests give us an insight into students' misconceptions of the problem. This change runs tests failing wheats through the chaff set. If any pass one or more chaffs, it gives us enough information to provide a hint.
Adding support for 'hints.json' to be added to the template to allow for hints to be displayed.